+++ ITS.CORE/Services/IIsIsciService.cs
@@ -11,7 +11,7 @@ namespace ITS.CORE.Services
 	public interface IIsIsciService
 	{
 		public Task<IEnumerable<CreateIsIsciDto>> AddIsIsciAsync(IEnumerable<CreateIsIsciDto> IsIsci);
-		public int ToplamGunAsync(int isciId, bool durumu);
+		public double ToplamGunAsync(int isciId, bool durumu);
 		public IEnumerable<CreateIsciBilgiDto> GunAllAsync(int isciId, bool durumu);
 		public Task<IEnumerable<IsIsci>> GetIsIsciAsync();
